requirements.txt
streamlit
requests
beautifulsoup4
selenium
gspread
oauth2client
pushbullet.py
captcha
